% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfo.R
\docType{package}
\name{dfo}
\alias{dfo}
\title{dfo: Handle Dept. Fisheries & Oceans Data Formats}
\description{
The \code{dfo} package provides functions for working with
Oceanographic datasets provided by the Department of
Fisheries and Oceans, Canada.
}
\details{
This package supplies functions for reading DFO data, returning
objects that are based on the \code{oce} base class.  Because
of this, the return values can be easily summarized, plotted,
and otherwise manipulated easily with \code{oce} functions.
This is illustrated within examples provided in
the documentation for the various \code{dfo} functions.
}
