% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ctd.ios.R
\name{read.ctd.ios.netcdf}
\alias{read.ctd.ios.netcdf}
\title{Read a CTD file in an IOS netCDF format}
\usage{
read.ctd.ios.netcdf(filename, missingValue = NULL, debug = 0)
}
\arguments{
\item{filename}{character value specifying the file name. This value must end
in `".nc"`.}

\item{missingValue}{numeric value that is to be converted to `NA`. If this is `NULL`,
the default, then an attempt to infer a reasonable value is made by
examining information stored in the file.}

\item{debug}{integer value controlling the printing of information
that may help in debugging problems. The default, 0, means to work
silently; positive values mean that some information should be printed.}
}
\value{
[read.ctd.ios()] returns an `oce` object of class `ctd`.
}
\description{
`read.ctd.ios.netcdf()` reads CTD data in a netCDF format used by
the Institution of Ocean Sciences, of the Department of Fisheries
and Oceans, Canada.  It starts by using [oce::read.netcdf()] to
decode the data.  Then, it sets the class to [oce::ctd-class],
after which it creates new variables in the `oce` naming
convention, with
`PRESPR01` becoming `pressure`,
`TEMPS901` becoming `temperature`,
`PSALST01` becoming `salinity`,
`DOXYZZ01` becoming oxygen in ml/l, and
`DOXMZZ01` becoming oxygen2 in mu*mol/kg.
}
\examples{
tempFile <- tempfile(fileext=".nc")
url <- "https://data.cioospacific.ca/erddap/files/IOS_CTD_Profiles/2007/2007-019-0055.ctd.nc"
download.file(url, tempFile)
ctd <- read.ctd.ios(tempFile)
summary(ctd)
# Note that using eos="gsw" in the following plot call will
# cause an error with oce version 1.3.0 (on CRAN as of Feb 2021),
# but it works with the updated oce, installed using
#     remotes::install_github("dankelley/oce", ref="develop")
plot(ctd, eos="unesco")
unlink(tempFile)

}
\references{
\url{https://data.cioospacific.ca/erddap/files/IOS_CTD_Profiles}
}
\author{
Dan Kelley
}
