% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ctd.ios.R
\name{read.ctd.ios}
\alias{read.ctd.ios}
\title{Read an IOS-formatted CTD file}
\usage{
read.ctd.ios(filename, missingValue = NULL, debug = 0)
}
\arguments{
\item{filename}{character value specifying the file name}

\item{missingValue}{numeric value that is to be converted to `NA`. If this is `NULL`,
the default, then the value is determined from the file, in the `CHANNEL DETAIL`
table within the `FILE` block.}

\item{debug}{integer value controlling the printing of information
that may help in debugging problems. The default, 0, means to work
silently; positive values mean that some information should be printed.}
}
\value{
[read.ctd.ios()] returns an `oce` object of class `ctd`.
}
\description{
This is a preliminary version of a function intended to read a data-file format
employed at the Institution of Ocean Sciences, of the Department of Fisheries
and Oceans, Canada.  It requires that the `oce` package be installed.
}
\section{Development note}{

1. Add lines to nameMaker() and unitMaker() as required, to handle
   column names that are not yet handled.
}

\examples{
file <- system.file("extdata", "2007-019-055.ctd", package="dfo")
ctd <- read.ctd.ios(file)
summary(ctd)
plot(ctd)

}
\references{
\url{https://catalogue.cioos.ca/dataset/ios_ctd_profiles}
}
\author{
Dan Kelley
}
