% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ctd.ios.R
\name{read.ctd.ios.text}
\alias{read.ctd.ios.text}
\title{Read a CTD file in an IOS text format}
\usage{
read.ctd.ios.text(filename, missingValue = NULL, debug = 0)
}
\arguments{
\item{filename}{character value specifying the file name. This must end
in \code{".ctd"}.}

\item{missingValue}{numeric value that is to be converted to \code{NA}. If this is \code{NULL},
the default, then an attempt to infer a reasonable value is made by
examining information stored in the file.}

\item{debug}{integer value controlling the printing of information
that may help in debugging problems. The default, 0, means to work
silently; positive values mean that some information should be printed.}
}
\value{
\code{\link[=read.ctd.ios.text]{read.ctd.ios.text()}} returns an \code{oce} object of class \code{ctd}.
}
\description{
\code{read.ctd.ios.text()} reads CTD data in a text format used by
the Institution of Ocean Sciences, of the Department of Fisheries
and Oceans, Canada.  Lacking a document from IOS describing
the format, the code is written based on inspection of
a particular file, and for this reason, it cannot be trusted
deeply.
}
\section{Development note}{

\enumerate{
\item Add lines to nameMaker() and unitMaker() as required, to handle
column names that are not yet handled.
}
}

\examples{
# File provided to author by a colleague.
file <- system.file("extdata", "2007-019-055.ctd", package="dfo")
ctd <- read.ctd.ios(file)
summary(ctd)
plot(ctd)

}
\references{
\url{https://catalogue.cioos.ca/dataset/ios_ctd_profiles}
}
\author{
Dan Kelley
}
