% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ctd.ios.R
\name{read.ctd.ios}
\alias{read.ctd.ios}
\title{Read an IOS-formatted CTD file}
\usage{
read.ctd.ios(filename, missingValue = NULL, debug = 0)
}
\arguments{
\item{filename}{character value specifying the file name.}

\item{missingValue}{numeric value that is to be converted to \code{NA}. If this is \code{NULL},
the default, then an attempt to infer a reasonable value is made by
examining information stored in the file.}

\item{debug}{integer value controlling the printing of information
that may help in debugging problems. The default, 0, means to work
silently; positive values mean that some information should be printed.}
}
\value{
\code{\link[=read.ctd.ios]{read.ctd.ios()}} returns an \code{oce} object of class \code{ctd}.
}
\description{
\code{read.ctd.ios()} reads CTD data in either of two formats used by
the Institution of Ocean Sciences, of the Department of Fisheries
and Oceans, Canada.  The formats are inferred from the value of the
\code{filename} argument.  If that value ends in \code{".ctd"}, then \code{\link[=read.ctd.ios.text]{read.ctd.ios.text()}} is
called to read the data.  If it ends in \code{".nc"}, then \code{\link[=read.ctd.ios.netcdf]{read.ctd.ios.netcdf()}}
is called. (No other endings are handled.)
}
\section{Development note}{

\enumerate{
\item Add lines to nameMaker() and unitMaker() as required, to handle
column names that are not yet handled.
}
}

\references{
\url{https://catalogue.cioos.ca/dataset/ios_ctd_profiles}
}
\author{
Dan Kelley
}
